+++ src/Exception/ServerMessageThrower.php
@@ -98,10 +98,11 @@ class ServerMessageThrower
             // "Your account has been disabled for violating our terms".
             '/account(.*?)disabled(.*?)violating/',
         ],
-        'SentryBlockException'                  => ['sentry_block'],
-        'Checkpoint\ScrapingWarningException'   => ['scraping_warning'],
-        'Checkpoint\UFACBloksException'         => ['ufac_www_bloks'],
-        'InvalidUserException'                  => [
+        'SentryBlockException'                      => ['sentry_block'],
+        'Checkpoint\ScrapingWarningException'       => ['scraping_warning'],
+        'Checkpoint\ReviewLinkedAccountsException'  => ['review_linked_accounts'],
+        'Checkpoint\UFACBloksException'             => ['ufac_www_bloks'],
+        'InvalidUserException'                      => [
             'invalid_user', // error_type
         ],
         'Checkpoint\ChangePasswordException'                 => ['/reset(.*?)password/', 'set_new_password', 'change_password'],